refactor(obj)!: rename `ObjMesh` to `RawObjMesh`
BREAKING CHANGE: Renamed `ObjMesh` structure to `RawObjMesh`.
Additionally rename `free_obj` to `free_obj_model` and `parse_obj_model`
to `parse_raw_obj`. This was needed to separate the `RawObjMesh`
structure from the GPU-optimized `ObjMesh` structure that will have a
monolithic array of vertices and an indices list. This affects all
references to the `ObjMesh` structure and `free_obj` and
`parse_obj_model` functions, so dependent code will not build until it
is updated accordingly.